+++ entrypoints/sidepanel/components/Chat/Messages/Assistant.vue
@@ -84,20 +84,27 @@
 </template>
 
 <script setup lang="ts">
-import { ref } from 'vue'
+import { computed, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
 
 import IconReasoningFinished from '@/assets/icons/reasoning-finished.svg?component'
 import IconWarning from '@/assets/icons/warning-circle.svg?component'
 import Loading from '@/components/Loading.vue'
 import Text from '@/components/ui/Text.vue'
-import { AssistantMessageV1 } from '@/utils/tab-store/history'
+import { AgentMessageV1, AssistantMessageV1 } from '@/types/chat'
 
 import MarkdownViewer from '../../../../../components/MarkdownViewer.vue'
-defineProps<{
-  message: AssistantMessageV1
+const props = defineProps<{
+  message: AssistantMessageV1 | AgentMessageV1
 }>()
 
+const message = computed(() => {
+  return {
+    ...props.message,
+    content: props.message.content.trim(),
+    reasoning: props.message.reasoning?.trim(),
+  }
+})
 const { t } = useI18n()
 const expanded = ref(false)
 </script>