+++ entrypoints/sidepanel/utils/chat/side-effects.ts
@@ -1,8 +1,8 @@
 import { effectScope, watch } from 'vue'
 
+import { ActionMessageV1 } from '@/types/chat'
 import { useGlobalI18n } from '@/utils/i18n'
 import { lazyInitialize } from '@/utils/memo'
-import { ActionMessageV1 } from '@/utils/tab-store/history'
 import { getUserConfig } from '@/utils/user-config'
 
 import { Chat } from './chat'