+++ CHANGELOG.md
@@ -1,6 +1,46 @@
 # Changelog
 
 
+## v1.7.0-beta.6
+
+[compare changes](https://github.com/NativeMindBrowser/NativeMindExtension/compare/v1.7.0-beta.5...v1.7.0-beta.6)
+
+### 🚀 Enhancements
+
+- **agent:** Add methods to persist and check current tab in AgentStorage ([510c90a](https://github.com/NativeMindBrowser/NativeMindExtension/commit/510c90a))
+
+### ❤️ Contributors
+
+- Tony Hu ([@tonyhu-012](http://github.com/tonyhu-012))
+
+## v1.7.0-beta.5
+
+[compare changes](https://github.com/NativeMindBrowser/NativeMindExtension/compare/v1.7.0-beta.4...v1.7.0-beta.5)
+
+### 🚀 Enhancements
+
+- **agent:** Improve environment info handling ([314fd98](https://github.com/NativeMindBrowser/NativeMindExtension/commit/314fd98))
+
+### ❤️ Contributors
+
+- Tony Hu ([@tonyhu-012](http://github.com/tonyhu-012))
+
+## v1.7.0-beta.4
+
+[compare changes](https://github.com/NativeMindBrowser/NativeMindExtension/compare/v1.7.0-beta.3...v1.7.0-beta.4)
+
+### 🚀 Enhancements
+
+- **agent:** Enhance agent tool execution and environment details handling ([9ec4df0](https://github.com/NativeMindBrowser/NativeMindExtension/commit/9ec4df0))
+
+### 🩹 Fixes
+
+- **test:** Fix chat prompt unit test ([532870a](https://github.com/NativeMindBrowser/NativeMindExtension/commit/532870a))
+
+### ❤️ Contributors
+
+- Tony Hu ([@tonyhu-012](http://github.com/tonyhu-012))
+
 ## v1.7.0-beta.3
 
 [compare changes](https://github.com/NativeMindBrowser/NativeMindExtension/compare/v1.6.1...v1.7.0-beta.3)